+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.0.1
 - fix english
+- bug fix ordering
 # 1.0.0
 - Pinning keybind is shift.
 - Pinning objects are in a json