+++ changelog.md
@@ -1,6 +1,7 @@
 # 1.0.1
 - fix english
 - bug fix ordering
+- new icon
 # 1.0.0
 - Pinning keybind is shift.
 - Pinning objects are in a json